Clean up epics and ensure correct order of actions
Epics were in a very prototype like stage before this commit.

With this commit we ensure that the state is prepared in the proper order so that the different actions can load their components in a balanced way between being independent from other components and having to call for the same information way to often.

Add title to SimpleProvenanceContract and adapt IT migration.