Add simple provenance contract and necessary libraries
In this commit we introduced a basic implementation of our provenance contract, focusing on the linking concept. This prototype is not complete and can not yet hold actual provenance information.

Note: After experiments we recognized that the test mocks have to be in the contracts directory otherwise truffle will not be able to start single tests but only the whole test suite.