Add contract deployment feature to react client
To achieve this it was necessary to create a own build step for truffle an build custom build outputs which contain the linked byte code of the contract. The byte code is build and linked on per network base. Thus it is necessary to migrate and build the truffle project when running on a new test chain before being able to deploy with the client.